Move AlternateName model to models
