+++ alternate_names.go
@@ -3,26 +3,16 @@ package geonames
 import (
 	"log"
 	"strconv"
+
+	"github.com/remizovm/geonames/models"
 )
 
 const alternateNamesURL = `alternateNames.zip`
 
-// AlternateName represents a single feature's alternate name
-type AlternateName struct {
-	ID              int    // alternateNameId   : the id of this alternate name, int
-	GeonameID       int    // geonameid         : geonameId referring to id in table 'geoname', int
-	IsoLanguage     string // isolanguage       : iso 639 language code 2- or 3-characters; 4-characters 'post' for postal codes and 'iata','icao' and faac for airport codes, fr_1793 for French Revolution names,  abbr for abbreviation, link for a website, varchar(7)
-	Name            string // alternate name    : alternate name or name variant, varchar(200)
-	IsPreferredName bool   // isPreferredName   : '1', if this alternate name is an official/preferred name
-	IsShortName     bool   // isShortName       : '1', if this is a short name like 'California' for 'State of California'
-	IsColloquial    bool   // isColloquial      : '1', if this alternate name is a colloquial or slang term
-	IsHistoric      bool   // isHistoric        : '1', if this alternate name is historic and was used in the past
-}
-
 // AlternateNames returns alternate names for all features available
-func (c *Client) AlternateNames() ([]*AlternateName, error) {
+func (c *Client) AlternateNames() ([]*models.AlternateName, error) {
 	var err error
-	var result []*AlternateName
+	var result []*models.AlternateName
 
 	zipped, err := httpGet(geonamesURL + alternateNamesURL)
 	if err != nil {
@@ -59,7 +49,7 @@ func (c *Client) AlternateNames() ([]*AlternateName, error) {
 			return true
 		}
 
-		result = append(result, &AlternateName{
+		result = append(result, &models.AlternateName{
 			ID:              id,
 			GeonameID:       geonameID,
 			IsoLanguage:     string(raw[2]),
